requirements.txt
cryptography==3.2.1
matplotlib==3.3.3
numpy==1.19.4
pandas==1.1.4
Pillow==8.0.1
PsychoPy==2020.2.8
wheel==0.35.1
setuptools==47.1.0
wesanderson==0.0.2